Adding a new function
Created a new function to list all participants and updated the menu to include this option.

Explanation of the Update:

New Function list_participants():
This function reads the secret_santa.csv file and lists all participants with a number to distinguish each participant.

Updated Menu:
The menu now includes a new option List Participants which calls the list_participants() function.